requirements.txt
torch
transformers
flask
requests
Flask-Cors
torchvision
Pillow